243: move Hash
