+++ src/rpc_client/eth_rpc/mod.rs
@@ -39,51 +39,6 @@ const HTTP_MAX_SIZE: u64 = 2 * 1024 * 1024;
 
 pub const MAX_PAYLOAD_SIZE: u64 = HTTP_MAX_SIZE - HEADER_SIZE_LIMIT;
 
-#[derive(Clone, Copy, Deserialize, Serialize, PartialEq, Eq, Hash, Ord, PartialOrd)]
-pub struct Hash(#[serde(with = "ic_ethereum_types::serde_data")] pub [u8; 32]);
-
-impl Debug for Hash {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{:x}", self)
-    }
-}
-
-impl Display for Hash {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{:x}", self)
-    }
-}
-
-impl LowerHex for Hash {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "0x{}", hex::encode(self.0))
-    }
-}
-
-impl UpperHex for Hash {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "0x{}", hex::encode_upper(self.0))
-    }
-}
-
-impl std::str::FromStr for Hash {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        if !s.starts_with("0x") {
-            return Err("Ethereum hash doesn't start with 0x".to_string());
-        }
-        let mut bytes = [0u8; 32];
-        hex::decode_to_slice(&s[2..], &mut bytes)
-            .map_err(|e| format!("failed to decode hash from hex: {}", e))?;
-        Ok(Self(bytes))
-    }
-}
-
-impl HttpResponsePayload for Hash {}
-
-impl HttpResponsePayload for Wei {}
-
 /// An envelope for all JSON-RPC requests.
 #[derive(Clone, Serialize, Deserialize)]
 pub struct JsonRpcRequest<T> {
@@ -268,6 +223,8 @@ pub trait HttpResponsePayload {
 impl<T: HttpResponsePayload> HttpResponsePayload for Option<T> {}
 
 impl HttpResponsePayload for TransactionCount {}
+
+impl HttpResponsePayload for Wei {}
 
 /// Calls a JSON-RPC method on an Ethereum node at the specified URL.
 pub async fn call<I, O>(
