+++ src/rpc_client/json/mod.rs
@@ -3,6 +3,7 @@
 use std::fmt::{Debug, Display, Formatter, LowerHex, UpperHex};
 use candid::Deserialize;
 use serde::Serialize;
+use crate::rpc_client::eth_rpc::HttpResponsePayload;
 
 pub mod requests;
 pub mod responses;
@@ -54,3 +55,46 @@ impl UpperHex for FixedSizeData {
         write!(f, "0x{}", hex::encode_upper(self.0))
     }
 }
+
+#[derive(Clone, Copy, Deserialize, Serialize, PartialEq, Eq, Hash, Ord, PartialOrd)]
+pub struct Hash(#[serde(with = "ic_ethereum_types::serde_data")] pub [u8; 32]);
+
+impl Debug for Hash {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{:x}", self)
+    }
+}
+
+impl Display for Hash {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{:x}", self)
+    }
+}
+
+impl LowerHex for Hash {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "0x{}", hex::encode(self.0))
+    }
+}
+
+impl UpperHex for Hash {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "0x{}", hex::encode_upper(self.0))
+    }
+}
+
+impl std::str::FromStr for Hash {
+    type Err = String;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        if !s.starts_with("0x") {
+            return Err("Ethereum hash doesn't start with 0x".to_string());
+        }
+        let mut bytes = [0u8; 32];
+        hex::decode_to_slice(&s[2..], &mut bytes)
+            .map_err(|e| format!("failed to decode hash from hex: {}", e))?;
+        Ok(Self(bytes))
+    }
+}
+
+impl HttpResponsePayload for Hash {}
